server/requirements.txt
fastapi[standard]==0.120.0
uvicorn[standard]==0.38.0
python-dotenv==1.0.1
groq==0.33.0
langchain-community==0.4.0
langchain-text-splitters==1.0.0
pypdf==6.1.3
python-multipart==0.0.20
tiktoken==0.12.0
langchain==1.1.0